Stabilize PythonInterpreter __eq__.
The prior check used the binary's stat which was unstable across
variations of mounts atime setting on different machines. If the
partition the python binary lived on had atime enabled (default in
OSX so an important case), then equality checks would ~always fail
for even the same binary.

This change switches to using binary real path + its reporter
version  info which should be all we care about.  It's true that
the binary might be replaced by a new one of the same version at
the same path but compiled with different options though, so we
still could do more and instead use binary real path + a content
hash of the binary if needed.