Bytecode compile using the correct interpreter.
Previously a PEXBuilder with a custom interpeter would not use it to
bytecode compile sources when building a pex.  This could lead to errors
when the current interpreter and the target interpreter spanned breaking
language changes.

Kill the CodeMarshaller which has only its marshalling half used in
favor of using py_compile from the stdlib in the new
interpreter-sensitive Compiler class.

Additionally, actually compile all PEXBuilder source including any
generated __init__.pys, the generated __main__.py and everything in the
_pex/ bootstrap.