Make it possible to use an alternate CA chain for pex
Pex is amazing in that it actually allows the end user to validate the
SSL when items are being obtained from an additional pypi repository.

Unfortunately, this validation will only occur for the authorities that
are specified in certifi. This limits two important use cases for
internal and none standard pypi servers:

* There is no method to use an alternate certificate authority chain,
  such as, for instance, is presented from a custom CA root
* There is no method to use client certificates, which can be an
  alternative to username and password.

This commit seeks to redress this, by allowing both of these options to
be specified, with configuration that mirrors that of pip.