Fix infinite recursion when ~/.pantsrc PEX_PYTHON points at a symlink
When ~/.pexrc sets PEX_PYTHON to a symlink, pex goes into an ~infinite loop.  On each iteration, it checks whether target python == current python, and if not, it reexecs as target python.  It previously called os.path.realpath on current but not on target, so for a symlink, they would never be equal.

Testing Done:
https://travis-ci.org/pantsbuild/pex/builds/91981334

Bugs closed: 182

Reviewed at https://rbcommons.com/s/twitter/r/3149/