Updates interpreter path parsing to handle Jenkins
When building a pex on Jenkins inside an auto-generated virtualenv, I
keep ending up with interpreter names like
CPython-2.7.3.tmp.ba3a2c39daab46ef978ba07fe335079d. If we simply split
on dots to get version number, we'll end up with an invalid version
number. To avoid this, we now just take the first 3 items when splitting
on a dot. This allows pex building for me on Jenkins.