Ensure `packaging` dependency is self-contained. (#355)
This dependency was not included in the pex `.bootstrap/` previously
leading to issues with use of pex as a library. We use the vendored
version from setuptools for now - which it leaks through its API. Notes
are added in the right spots to deal with setuptools de-vendoring of
`packaging` in its newest releases when we upgrade setuptools next.