Fix corner case in cached dependency resolution (#362)
Fixes #178

The error is as follows:

Two resolvables, one pinned, the other not.
First pex run with clean cache works.
Change the pinned version.
Second pex run fails with Unsatisfiable.
The problem is that the non-pinned resolvable locates the cached dep
and only returns it. Since that version doesn't match the pinned
version, we fail. This fixes that behavior by returning both the
cached version and all uncached pypi packages for unpinned
resolvables.

We change the pinned version to test the use case where you have a
cached version of a dep and then you transitively pull in another
requirement that either has a tighter bound or a pin on a version not
found in the cache.