Adds `-m` option to saved .pex file interpreter emulation
This allows you to invoke a module from a saved PEX file with
`app.pex -m <module>`, in addition to invoking a script with
`app.pex <file>` and a REPL with `app.pex`, as could already be
done previously.

This makes a saved PEX file more suitable as a basic replacement
for the `python` binary on platforms that allow you to do that but
that don't have good support for Python dependency management
themselves, such as PySpark.