Add interpreter constraints option and use constraints to search for compatible interpreters at exec time (#427)
Problem:
PEX does not support interpreter compatibility ranges for selecting an interpreter to build/run a PEX with. Twitter is aiming at killing off PEX_PYTHON overriding and replacing with a PEX_PYTHON_PATH in /etc/pexrc, which will contain a path to directories containing canonical interpreters for searching against at pex bootstrap re-exec time. Interpreter constraints on a PEX should inform the searching of PEX_PYTHON_PATH (if present in the env) and error out if no compatible interpreter can be found, replacing the searching of $PATH as a fallback.

Solution:
Add an interpreter constraints option to pex CLI and write these constraints to PEX-INFO metadata. Add an env variable to variables.py to support the setting of PEX_PYTHON_PATH. When the bootstrapper pex re-execs at runtime, read constraints from PEX-INFO and attempt to resolve an interpreter to exec with from the PPP. If no interpreter can be found in PPP, error out indicating that no match can be found.

Result:
PEXes will be able to be built and executed with specified interpreter constraints. Users will be able to define and control interpreter selection in managed development environments by setting the PEX_PYTHON_PATH so that PEXes are executed with supported interpreters only.