Change PEX re-exec variable from ENV to os.environ (#441)
Currently, if an application packaged as a pex builds and runs a product pex as part of its functionality (i.e. a Pants pex that builds and runs pexes), the interpreter selection logic in a product pex will short-circuit because the flag telling it to do so has already been set in the initial bootstrapping logic of the main app's pex. Consequently, when the main app runs a product pex as part of its functionality, the interpreter selection in accordance with a pexrc for the product pex is skipped because the environment (os.environ) now contains the SHOULD_EXIT_BOOTSTRAP_REEXEC flag set to "True" from the bootstrap phase of the main app's pex.

This is because SHOULD_EXIT_BOOTSTRAP_REEXEC (the flag governing whether or not to re-exec) was being set in an ENV object variable, which is composed into the environment of the re-execution call os.execve via the env parameter.

The global instance of ENV is unique to each isolated Pex environment, but for this use case we need environment data that can be shared across Pex environments. This changes in this PR switch the implementation of SHOULD_EXIT_BOOTSTRAP_REEXEC from ENV to os.environ, where changes to this variable will be visible across different pex environments running from the same Python process.