Support `-` when running as an interpreter.
It's surprising when using a pex in python interpreter mode that `-` is
not recognized as a program coming from stdin as regular python
interpreter would. Add support for this with a test.

The surprise of this missing feature was rightly identified in the
context of https://github.com/pantsbuild/pants/pull/6275.