Fix PEXBuilder.clone. (#575)
Previously clone copied distributions twice. One copy was made by the
underlying `Chroot`, and then another was manually performed. The
manual copy was only needed to update the internal `_distributions`
set, so just do that.

A test that failed before the fix is added to ensure distributions with
unwriteable files can be cloned.

Fixes #570