Pex exits with correct code when using entrypoint
Currently the exit code from an entrypoint is swallowed, so running pex
always exits 0.

Tested with:
```
tox -e py27-package
dist/pex27 -e pytest:main pytest -D /path/to/failing/test/dir
```

before this PR, this exits 0, after it exits 1.

I'm not sure how to properly integration test this, as it is in the
outer-most wrapper layer of the pex binary, so looks like it's bypassed
by the way we tend to test things?