Follow PEP 425 suggestions on distribution preference.
PEP 425 describes how a single interpreter on a platform can support many
tags and a single package might have many distributions to satisfy those
supported tags [0]. It describes a tie breaking mechanism:

> Sometimes there will be more than one supported built distribution for a
> particular version of a package. For example, a packager could release a
> package tagged cp33-abi3-linux_x86_64 that contains an optional C extension
> and the same distribution tagged py3-none-any that does not. The index of the
> tag in the supported tags list breaks the tie, and the package with the C
> extension is installed in preference to the package without because that tag
> appears first in the list.

pex lacks this tie breaking logic currently and it is the root cause of
issue #639.

To address this issue, I have enhanced the resolver's sorter to deprioritize
packages that support the 'any' platform.

[0]: https://www.python.org/dev/peps/pep-0425/#id1