Restore allowing all discoverable interpreters when no constraints
Several tests were failing that they could not find a compatibile interpreter for the empty constraints [].

We did not originally hit this error because the builtin all() will return True if the iterable is empty. So, we were calling all(matches(filt) for filt in []), which returned True. Now that we use any(), it would return False.

Instead, we should only call matched_interpreters() if there are any constraints to begin with.