Correct .travis.yml tox syntax to work for 3.8.1+. (#693)
The tox 3.8.1 release requires `-v` and `-e <env>` be passed as
seperate arguments where, before, `-ve <env>` was accepted. Since we
float our tox dependency in CI to match our lack of a bound on
developer tox installs, update to the new CLI syntax since it is
compatible with old tox too.