Use all compatible versions when calculating tags. (#692)
Previously, we'd pass a single version to `get_supported` when we knew
one which would lead to a single tag being generated for the abi3 case
instead of one per previous minor version.

As part of this change, we need a more modern version of the
pep425tags.py we had copied from pip that allows passing multiple
versions to `get_supported`. Ideally, we'd use the version embedded in
our vendored setuptools code and this is tracked by #696. Extra code
had been added to pep425tags to implement `get_supported_for_any_abi`
and handle manylinux. This is now inlined in platforms.py, which was
the only user,  the door to using undisturbed copies in #696.

Fixes #539