When iterating over Zipfiles, always use the Unix file separator to fix a Windows issue (#638)
Pex doesn't officially support windows, but there seems to be a regression in pex 1.6.0 with windows with respect to the way the vendoring works. Specifically, the regex for iterating the zip path in the vendoring uses the os separator, which has two problems on windows (where the separator is `\`):

1. `\` is an escape sequence, so unless the separator is added twice in the string, it uses it to escape the character immediately after the separator, causing the regex compilation to fail.
1. The path in the zip file uses unix separators (`/`), so using the os separator is actually wrong.

This PR addresses this by updating the regex to replace windows separators (`\`) in the prefix var with the unix separator.