Introduce new `--no-compile` flag to not include .pyc in built pex due to its non-determinism (#718)
### Problem
Until Python 3.7 via https://www.python.org/dev/peps/pep-0552/, `.pyc` files were never reproducible because they included the timestamp. See the discussion at https://github.com/pantsbuild/pex/issues/716#issuecomment-487821354.

We are aiming to ensure reproducible builds a la https://github.com/pantsbuild/pex/issues/716 so that we can safely remote PEXes in CI. Reproducible builds also play an important role in security, per https://reproducible-builds.org.

### Solution
Allow users to toggle on and off `.pyc` files via the new `--compile` / `--no-compile` flag.

We currently default to including `.pyc`, because this has been the precedent for 9 years. In a future release, e.g. 1.70, we will change the default to not include `.pyc` so that builds are reproducible by default, but people can still opt into including the files if they'd like via the flag.

#### Alternative flag considered: `--reproducible`
Originally, we were going to use `--reproducible` to toggle on multiple behaviors like not including `.pyc` and using a hardcoded timestamp instead of system time. However, after deciding that in a future release we are going to default to reproducible, it doesn't make sense for people to ever call `--not-reproducible`, but it does make sense to opt out of specific behaviors like wanting to use `--include-pyc`. So, we use behavior-specific flags instead of a universal flag.

#### Alternative solution: ensure our own `.pyc` timestamp
It may be possible to still include `.pyc` _and_ have their timestamp be deterministic https://github.com/pantsbuild/pex/pull/718#issuecomment-488202220. This is not pursued for now because it is more complex than necessary, especially because currently we do not ship `.pyc` files if multiple platforms are used in the PEX. But it could be a followup PR if deemed worth it.

### Result
When `--no-compile` is set, PEXes will no longer include `.pyc`. This results in all of the [reproducible build acceptance tests](https://github.com/pantsbuild/pex/blob/master/tests/test_integration.py#L1378) passing the exploded pex portion of their tests, excluding the `bdist_wheel` test.

#### Impact on performance
Not using `.pyc` has a slight speedup of ~0.2 seconds to the _creation_ of simple PEXes.

* `time python -m pex -o normal.pex --compile` averaged around 0.49 seconds
* `time python -m pex -o repro.pex --no-compile` averaged around 0.31 seconds

Not using `.pyc` has a slight slowdown of ~0.06 seconds to the startup time of simple PEXes.

* `time ./normal.pex --version` averaged around 0.26 seconds
* `time ./repro.pex --version` averaged around 0.32 seconds

`python -m pex -m pydoc -o pydoc.pex` likewise ran 0.07 seconds slower when not including `.pyc`.

Note that the inclusion of `.pyc` only impacts startup time, not actual runtime performance. See https://stackoverflow.com/questions/2998215/if-python-is-interpreted-what-are-pyc-files for what `.pyc` files do / why they exist.