Revert making deterministic_datetime a constant
If it's a constant, theen it becomes very difficult to test that $SOURCE_DATE_EPOCH works because constants get evaluated at import time, and at import time os.environ does not have the env var set. It will ignore the test setting the value, because the constant never gets re-evaluated.