Drop overly verbose comment about SOURCE_DATE_EPOCH
We don't need to explain this here because readers of the code are unlikely to anticipate it being an issue, so are likely to get distracted more than anything. Instead, this PR description is updated to make the explanation if anyone is wondering "wait why didn't they do this?"

Also make the MS-DOS explanation more concise.