Fix PexInfo requirements using a non-deterministic data structure (#723)
## Problem
Even though `PexInfo` stores its requirements as an `OrderedSet`, when getting converted into a `list` in the `dump()` function, the requirements were losing their deterministic order for an unclear reason.

Specifically, the acceptance test diverged with `PEX-INFO`, with both files having different orders for their requirements. 

```
# pex1
"requirements": ["six==1.12.0", "cryptography==2.6.1", "cffi!=1.11.3,>=1.8", "asn1crypto>=0.21.0", "pycparser"]

# pex2
["six==1.12.0", "cryptography==2.6.1", "asn1crypto>=0.21.0", "cffi!=1.11.3,>=1.8", "pycparser"]
```

## Solution
Sort the requirements when being dumped to ensure we always have a consistent result.

We also make `interpreter_constraints`'s output always be sorted, as it is internally stored as a set so by definition is not deterministic in ordering.