Fix reproducible build test improperly using the -c flag and add a new test for -c flag (#725)
The `-c` acceptance test was failing not due to unreproducibility, but because the Pex was failing to be built in the first place. It simply needed to have the corresponding Python file be created to get working. Note that the failing test passed locally on macOS and an Ubuntu VM, but was failing in Travis—it now passes universally.

We also add a test to cover using `-c` from a 3rd party dependency, as we were originally only testing when building from source. Both are important cases to ensure reproducibility.