add delayed pkg_resources import fix from #713, with an integration test (#730)
This is a fix to the way we load pkg_resources, along with an integration test which fails on master since the setuptools vendoring in #624, but is fixed by the change in #713. This can be merged before that one to fix the failures induced in several 3rdparty packages, while #713 can be left to understanding and testing the separate but related problem that this change also happens to fix.