Add an --intransitive option.
If specified, only the explicitly provided requirements are
resolved.  Their transitive requirements are not.

Since pex checks that all requirements are satisfied, this means
that a complete transitive set must be provided explicitly in
order to use this flag.

This is useful if you have a lockfile from some other source
and want pex to just apply it.