Fix current platform handling.
Three aspects of current platform handling are fixed:

1. Ensure that `Platform.supported_tags` returns the full range of tags
when the platform is current. This uses the special no-arg form of
`pep425tags.get_supported` which consults the local interpreter to
return the broadest possible set of tags. Also ensure that we always
pass all compatible versions when the platform is not current to
generate the broadest possible set of tags for the foreign platform.

2. Ensure that the resolver interprets 'current' to mean any local
interpreter instead of just the current runtime interpreter. This is
faithful to prior behavior where a PEX built for the 'current' platform
would always work on the current machine.

3. Fix the PEX cli when run with a fully foreign set of `--platform`s.
Previously the attempt to warn the user would error on an invalid log
call that was missing a `V` argument.

Tests of all of the above are added and platform parsing is pushed up to
the CLI and resolver API layers for cleaner code under those layers and
earlier failures for invalid platform specifiers.