Cleanup resolver.
This extracts some calculations out of the main `resolve_distributions`
method and streamlines install request categorization to use a similar
form to build request categorization.

Also cleanup marker calculation with some consolidation of code as well
as elevation of that tail portion of the resolve to an officially
commented resolve stage.