Parallelize resolve. (#819)
The three major phases of a pex resolve: download, build and install,
are delegated to pip subprocesses. As such, we can easily parallelize
these operations only needing to take care of shared portions of the
filesystem the processes might mutate. In the end this is only relevant
to the build and install phases which are natural points to cache
results in a shared filesystem cache. The install phase is already
cached in this way (#815) and we add caching for the build phase as
well, both utilizing (posix in the end) guarantees around `os.rename`.

The atomic shared directory updates are achieved with `AtomicDirectory`
and the subprocess parallelization is acheived using the new `jobs`
module and request / response data object pairs to coordinate parallel
pip jobs.

Example speedups:

+ PEX build using prebuilt wheels (the #811 case):
  ```
  pex \
    --platform=manylinux1-x86_64-cp-35-m \
    --platform=manylinux1-x86_64-cp-36-m \
    --platform=manylinux1-x86_64-cp-37-m \
    --platform=macosx-10.9-x86_64-cp-35-m \
    --platform=macosx-10.9-x86_64-cp-36-m \
    --platform=macosx-10.9-x86_64-cp-37-m \
    numpy==1.17.4 \
    --python-shebang "/usr/bin/env python3" \
    -o numpy-pex1.6.2.pex
  ```
  pex version | cold      | warm
  ------------|-----------|----------
  1.6.12      | 0m41.313s | 0m20.759s
  2.0.2       | 0m53.236s | 0m27.217s
  HEAD        | 0m32.336s | 0m17.596s

+ PEX build including sdists (wheel builds):
  ```
  pex \
    pantsbuild.pants==1.22.0 \
    -o pantsbuild.pants.pex
  ```
  pex version | cold      | warm
  ------------|-----------|----------
  1.6.12      | 0m49.101s | 0m11.788s
  2.0.2       | 1m15.602s | 1m2.476s
  HEAD        | 0m47.174s | 0m15.309s

Fixes #811
Fixes #817
Fixes #818