Fix pex resolving for foreign platforms.
This change restores manylinux handling to the pre pex 2.0.0 status quo
while properly supporting manylinux-next via the new `--manylinux`
option that takes a manylinux platform upper bound to assume whenever
the user specified platform is linux. This upper bound now defaults to
manylinux2014.

The change also brings in a pip fix for abi3 handling for foreign
platforms and a test is added to track this continues to work.

Fixes #823