Re-enable pyenv interpreter tests under pypy.
It turned out the fundamental issue was a bad version of curl which was
actually hinted at by the dump originally included in #477. We've since
updated our linux image to bionic which happened to fix the issue moving
to a newer curl as a side-effect.

This change re-enables disabled tests as well as converting a test
skipped previously due to this blanket issue to a more specific,
legitimate skip. The `.travis.yml` is also cleaned up a bit.

The hint this might be worth pursing came from:
  https://github.com/unbit/uwsgi/issues/37#issuecomment-10295067

Fixes #477