Check expected pex invocation failure reason in tests.
In some tests we just checked that expected invocation failures
happened. These tests continued to pass when the failure reason
changed from the one we expected, to an unrelated one having to
do with manylinux tag incompatibilities.

This change makes the tests check the invocation's stderr for the
expected failure output.