Use one version of transitively vendored deps
This is kind of hacky. It probably shouldn't be a hard-coded list in
__main__.py.

It's also not guaranteed to work - it relies on us picking good versions
to vendor which are compatibile with all of the transitive Avendored
versions.

We also don't vendor all of the transitive deps, because the rest don't
seem to be used by our codepaths. This is a bit fragile.

Finally, the last rewrite, `pex.third_party.six` ->
`pex.vendor._vendored.six.six` shouldn't be necessary. If the approach
seems reasonable, I'll dig into why it's needed to work, and resolve
that.