Support for copying requirements from an existing pex. (#948)
This can greatly speed up repeatedly creating a pex when
no requirements have changed. A build tool (such as Pants)
can create a "requirements pex" that contains just a static
set of requirements, and build a final pex on top of that,
without having to re-run pip to resolve requirements.