use relative paths within wheel cache (#979)
When running pex within a chroot environment, the symlinks generated by pex/pex/resolver.py use the absolute path within the chroot in the path for the symlinks' targets. These symlinks are invalid paths to any program that accesses the wheel cache outside of the chroot. (This was seen within a remote execution environment with Buildfarm.)

Switch to using relative paths for the symlinks.

Fixes #978.