Ensure pip download dir is uncontended.
When pip encounters an existing file during any of its file operations
it prompts for an action if `--exists-action` has not been set. Pex was
supplying a download directory to pip that was degenerate for
interpreters with the same abi and same major and minor versions (e.g.:
2.7.6 and 2.7.18). We ensure the download directory is unique per
interpreter and further configure the `--exists-action` to abort to get
fail fast behavior and smoke out regressions or any remaining issues in
this vein.

Fixes #983