Use MyPy via type comments (#1032)
Note that we do not actually need to vendor `typing`, even though it's not provided by Python 2. Read the docstring in `pex/typing.py` to see how we use a similar approach to Pip so that the runtime does not use anything from `typing`, but MyPy is still able to work.

Once we require Python 3 to run, we can remove this facade and use the built-in `typing` without conditional imports.