Add type hints to most tests (#1036)
Now all tests have full coverage except for `test_environment.py`, `test_pex.py`, and `test_integration.py`, which are saved for a dedicated followup.

Why start with tests? These are the call-sites of much of our production code, and this will help to ensure that when we add types to our production code, the hints reflect reality. The tests also are "roots", which makes it easier to add hints to than adding hints to code with lots of dependees.

### `str` vs. `bytes` vs. `Text` vs. `AnyStr`

Python 2 lets you freely mix unicode and bytes :/ Pex uses this abundantly, in contrast to Pants's (awkward) use of `future` to ensure that Py2 always used the exact same type as Py3.

Originally, this PR added `from __future__ import unicode_literals`, but that was reverted to instead optimize for fewer code changes. PEX's Py2 support works well as is, and it's well tested. It's not worth the energy to do a Pants-style migration that ensures Py2 behaves like Py3.

Instead, these hints try to capture the truth, as follows:
* `bytes` -> when both Py2 and Py3 use `bytes`
* `str` -> when Py2 uses `bytes`, but Py3 uses unicode
* `Text` -> when both use unicode. This only really happens when we explicitly call `.decode('utf-8')`
* `AnyStr` -> when it can be either unicode or bytes; shorthand for `Union`

Once we require Py3 to run, `Text` and `AnyStr` will both be converted into simply `str` (i.e. unicode).