Add type hints to variables.py and platforms.py (#1042)
Unfortunately, so long as we require Python 2, we cannot use `namedtuple` robustly with MyPy because it does not support annotating the fields of a named tuple. So, we recreate namedtuple through our own conventional class. This can be fixed once we require Python 3 to run. See https://github.com/pantsbuild/pex/issues/1041.